fix: move test to new suite to avoid event server conflict
