+++ src/tests-rosetta/api.ts
@@ -23,7 +23,6 @@ import {
   SignedTokenTransferOptions,
   someCV,
   standardPrincipalCV,
-  stringAsciiCV,
   TransactionSigner,
   tupleCV,
   uintCV,
@@ -32,7 +31,7 @@ import {
 } from '@stacks/transactions';
 import * as BN from 'bn.js';
 import { StacksCoreRpcClient } from '../core-rpc/client';
-import { bufferToHexPrefixString, I32_MAX, timeout } from '../helpers';
+import { bufferToHexPrefixString, timeout } from '../helpers';
 import {
   RosettaConstructionCombineRequest,
   RosettaConstructionCombineResponse,
@@ -74,8 +73,6 @@ import {
 } from '../rosetta-helpers';
 import { makeSigHashPreSign, MessageSignature } from '@stacks/transactions';
 import { decodeBtcAddress } from '@stacks/stacking';
-import { TestBlockBuilder } from '../tests/test-helpers';
-import { createClarityValueArray } from '../p2p/tx';
 
 
 describe('Rosetta API', () => {
@@ -3293,135 +3290,6 @@ describe('Rosetta API', () => {
   })
 
   /* rosetta construction end */
-
-  test('block/transaction - contract_call contains parsed metadata', async () => {
-    const testContractAddr = 'ST27W5M8BRKA7C5MZE2R1S1F4XTPHFWFRNHA9M04Y.hello-world';
-    const contractJsonAbi = {
-      maps: [],
-      functions: [
-        {
-          args: [
-            { type: 'uint128', name: 'amount' },
-            { type: 'string-ascii', name: 'desc' },
-          ],
-          name: 'test-contract-fn',
-          access: 'public',
-          outputs: {
-            type: {
-              response: {
-                ok: 'uint128',
-                error: 'none',
-              },
-            },
-          },
-        },
-      ],
-      variables: [],
-      fungible_tokens: [],
-      non_fungible_tokens: [],
-    };
-
-    const block = await api.datastore.getCurrentBlock();
-    assert(block.found);
-    let currentHeight = block.result.block_height;
-    const block4 = new TestBlockBuilder({
-      block_height: ++currentHeight,
-      index_block_hash: '0x04',
-      parent_index_block_hash: block.result.index_block_hash,
-    })
-      .addTx({ tx_id: '0x1111' })
-      .addTxSmartContract({ contract_id: testContractAddr, abi: JSON.stringify(contractJsonAbi) })
-      .addTxContractLogEvent({ contract_identifier: testContractAddr })
-      .build();
-    await db.update(block4);
-    const block5 = new TestBlockBuilder({
-      block_height: ++currentHeight,
-      index_block_hash: '0x05',
-      parent_index_block_hash: '0x04',
-      block_hash: '0xf1f1',
-    })
-      .build();
-    const contractCall: DbTx = {
-      tx_id: '0x1112',
-      tx_index: 1,
-      anchor_mode: 3,
-      nonce: 0,
-      raw_tx: Buffer.alloc(0),
-      index_block_hash: block5.block.index_block_hash,
-      block_hash: block5.block.block_hash,
-      block_height: block5.block.block_height,
-      burn_block_time: block5.block.burn_block_time,
-      parent_burn_block_time: 1626122935,
-      type_id: DbTxTypeId.ContractCall,
-      status: 1,
-      raw_result: '0x0100000000000000000000000000000001', // u1
-      canonical: true,
-      microblock_canonical: true,
-      microblock_sequence: I32_MAX,
-      microblock_hash: '',
-      parent_index_block_hash: '',
-      parent_block_hash: '',
-      post_conditions: Buffer.from([0x01, 0xf5]),
-      fee_rate: 10n,
-      sponsored: false,
-      sponsor_address: undefined,
-      sender_address: testContractAddr,
-      origin_hash_mode: 1,
-      event_count: 5,
-      execution_cost_read_count: 0,
-      execution_cost_read_length: 0,
-      execution_cost_runtime: 0,
-      execution_cost_write_count: 0,
-      execution_cost_write_length: 0,
-      contract_call_contract_id: testContractAddr,
-      contract_call_function_name: 'test-contract-fn',
-      contract_call_function_args: createClarityValueArray(uintCV(123456), stringAsciiCV('hello')),
-      abi: JSON.stringify(contractJsonAbi),
-    };
-    block5.txs.push({
-      tx: contractCall,
-      stxLockEvents: [],
-      stxEvents: [],
-      ftEvents: [],
-      nftEvents: [],
-      contractLogEvents: [],
-      smartContracts: [],
-      names: [],
-      namespaces: [],
-    });
-    await db.update(block5);
-
-    const query1 = await supertest(api.server)
-      .post(`/rosetta/v1/block/transaction`)
-      .send({
-        network_identifier: { blockchain: 'stacks', network: 'testnet' },
-        block_identifier: { index: currentHeight, hash: '0xf1f1' },
-        transaction_identifier: { hash: '0x1112' },
-      });
-    expect(query1.status).toBe(200);
-    expect(query1.type).toBe('application/json');
-    const result = JSON.parse(query1.text);
-    expect(result.transaction_identifier.hash).toEqual('0x1112');
-    expect(result.operations[1].metadata).toEqual({
-      contract_id: 'ST27W5M8BRKA7C5MZE2R1S1F4XTPHFWFRNHA9M04Y.hello-world',
-      function_args: [
-        {
-          hex: '0x010000000000000000000000000001e240',
-          name: 'amount',
-          repr: 'u123456',
-          type: 'uint',
-        },
-        {
-          hex: '0x0d0000000568656c6c6f',
-          name: 'desc',
-          repr: '"hello"',
-          type: 'string-ascii',
-        },
-      ],
-      function_name: 'test-contract-fn',
-      function_signature: '(define-public (test-contract-fn (amount uint) (desc string-ascii)))',
-    });
-  });
 
   afterAll(async () => {
     await new Promise<void>(resolve => eventServer.close(() => resolve()));
